django/requirements.txt
# Django backend
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
gunicorn
asgiref
psycopg2-binary
django-cors-headers
sqlparse
requests
certifi

# Image / font pipeline
fonttools
Pillow~=10.4
fonttools~=4.55
opencv-contrib-python-headless==4.10.0.84